Add ideavimrc

+++ install.sh
@@ -3,9 +3,10 @@
 set -ex
 
 pwd=`pwd`
-files=(alacritty.yml vimrc zshrc tmux.conf p10k.zsh)
 
-for f in "${files[@]}"; do
-    rm -rf $HOME/.$f
-    ln -s $pwd/$f $HOME/.$f
+for f in $(ls); do
+    if [[ $f != "install.sh" ]]; then
+        rm -rf $HOME/.$f
+        ln -s $pwd/$f $HOME/.$f
+    fi
 done
